refactor: 불필요한 인자 전달 삭제
